Add other actions. Change action calc result from int to float

+++ src/Actions/ActionInterface.php
@@ -12,7 +12,7 @@ interface ActionInterface
      *
      * @throws ActionException
      *
-     * @return int
+     * @return float
      */
-    public function calc(int $a, int $b): int;
+    public function calc(int $a, int $b): float;
 }
